requirements.txt
pymupdf
sentence-transformers
faiss-cpu
numpy
tiktoken
openai
